TS-4751: Prune cached headers before merging the updated response.
When we update a cached response with a new server response, make
sure to delete any cache-related beforehand. This is needed to
handle the situation where an origin returns Age on an initial
response but not on subsequent responses. We should not preserve
the initial Age in this case.

(cherry picked from commit 301561ce9bc73d19a18809d987da1827affdda5e)