stale_while_revalidate: Continue revalidation while stale on net thread.
Swapping the revalidation to a task thread causes a number of lookups to crash ATS,
as they’re designed to require net thread resources. Letting the revalidation continue
on a net thread gives it access to the pools and other resources it requires. This fixes
segfaults that occur when stale_while_revalidate is loaded.