Added detection for excessively old state machines.
There exist circumstances, causes hitherto unknown, that allow the state
machines to get into an infinite looping state. This creates excessive load and
causes serious problems for both mids and origins. This doesn't fix the loop,
but it terminates state machines at 100 transitions, which is hopefully more
than an order of magnitude larger than expected transition counts.

In addition, it enables debugging for state machines that exceed 20
transitions, which will hopefully give us insight into what is causing the
trouble, when it occurs.